requirement.txt
streamlit
azure-storage-blob
werkzeug
